backend/requirements.txt
flask
flask-sqlalchemy
psycopg2
openai
python-dotenv